### 2019-08-06 #1630 @whedon
[PRE REVIEW]: virtualNicheR: generating virtual fundamental and realised niches for use in virtual ecology experiments
**Submitting author:** @tretherington (<a href="http://orcid.org/0000-0002-3187-075X">Thomas R Etherington</a>)
**Repository:** <a href="https://github.com/manaakiwhenua/virtualNicheR" target ="_blank">https://github.com/manaakiwhenua/virtualNicheR</a>
**Version:** v1.0
