### 2019-08-06 #1630 @tretherington
Hi @arfon - and thank you for setting up the journal! :smile:

Based on a filter for "ecology" and "R" I would suggest @trashbirdecology, @ha0ye, or @ViniciusBRodrigues
