### 2019-08-19 #1630 @lsw5077
Hello,

Yes, I can review it.

-Lyndsie

On Mon, Aug 19, 2019 at 10:10 AM Arfon Smith <notifications@github.com>
wrote:

> 👋 @lsw5077 <https://github.com/lsw5077> - would you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
