### 2019-08-19 #1661 @whedon
[REVIEW]: virtualNicheR: generating virtual fundamental and realised niches for use in virtual ecology experiments
**Submitting author:** @tretherington (<a href="http://orcid.org/0000-0002-3187-075X">Thomas R Etherington</a>)
**Repository:** <a href="https://github.com/manaakiwhenua/virtualNicheR" target ="_blank">https://github.com/manaakiwhenua/virtualNicheR</a>
**Version:** v1.0
**Archive:** <a href="https://doi.org/10.7931/cae9-bt94" target="_blank">10.7931/cae9-bt94</a>
