### 2019-08-24 #1661 @ViniciusBRodrigues
Based on the review checklist, small modifications should be done. The **A statement of need** and **Community guidelines** could be improved. The use of the software could be very wide inside ecology, but you can specify (more clearly) the target audience.

There is no clear information about how to contribute to the software, report issues or seek support. Despite the users can do that through GitHub, you should make it clear and inform options for non-GitHub users, I think. :grin:
