### 2019-08-29 db92274
JOSS review changes
