+++ paper/paper.bib
@@ -31,6 +31,17 @@
    type = {Journal Article}
 }
 
+@article{elith-2009,
+   author = {Elith, J. and Leathwick, J. R.},
+   title = {Species distribution models: ecological explanation and prediction across space and time},
+   journal = {Annual Review of Ecology Evolution and Systematics},
+   volume = {40},
+   pages = {677-697},
+   DOI = {10.1146/annurev.ecolsys.110308.120159},
+   year = {2009},
+   type = {Journal Article}
+}
+
 @article{etherington-2015,
    author = {Etherington, T. R. and Holland, E. P. and O'Sullivan, D.},
    title = {NLMpy: a Python software package for the creation of neutral landscape models within a general numerical framework},
