+++ paper/paper.md
@@ -21,9 +21,9 @@ bibliography: paper.bib
 
 # Introduction
 
-Models of species-environment relationships (such as: habitat suitability models, species distribution models, resource selection functions, or ecological niche models) are an important part of ecological science.  The foundation of many of these models is the concept of fundamental and realised ecological niches that delineate the combinations of abiotic and biotic conditions under which a species can persist indefinitely [@hutchinson-1957].
+Models of species-environment relationships (such as: habitat suitability models, species distribution models, resource selection functions, or ecological niche models) are an important part of ecological science [@elith-2009].  The foundation of many of these models is the concept of fundamental and realised ecological niches that delineate the combinations of abiotic and biotic conditions under which a species can persist indefinitely [@hutchinson-1957].
 
-The ``virtualNicheR`` software we have developed creates virtual fundamental and realised niches for virtual, or imaginary, species.  The ability to generate virtual niches enables the virtual ecology experiments [@zurell-2010] that have been advocated for evaluating species-environment relationship models [@miller-2014].  There are a variety of existing software tools that aim to do this [@duan-2015; @leroy-2016; @qiao-2016], but none of the existing software differentiate between fundamental and realised niches, which is important given differences in possible complexity of shape [@blonder-2016].  The ``virtualNicheR`` software that we have developed is therefore unique in that it provides an integrated framework within which to generate both fundamental and realised niches.
+To evaluate how well species-environment models perform in different situations, modellers need the ability to generate virtual, or imaginary, niches with known properties.  There are a variety of existing software tools that aim to do this [@duan-2015; @leroy-2016; @qiao-2016], but none of the existing software differentiate between fundamental and realised niches, which is important given differences in possible complexity of shape [@blonder-2016].  The ``virtualNicheR`` software that we have developed is therefore unique in that it provides an integrated framework that is needed to generate both fundamental and realised virtual niches.  We envisage that the ability to generate integrated fundamental and realised virtual niches will enable modellers to improve the virtual ecology experiments [@zurell-2010] that have been advocated for evaluating species-environment relationship models [@miller-2014].
 
 # Virtual niche framework -- from sets to continuums
 
