+++ README.Rmd
@@ -16,12 +16,11 @@ knitr::opts_chunk$set(
 ```
 # virtualNicheR <img src="man/figures/logo.png" align="right" width="200" />
 
-`virtualNicheR` is an `R` package that generates virtual [fundamental and realised niches](https://en.wikipedia.org/wiki/Ecological_niche)
-that can be used as the basis for virtual ecology experiments examining questions around ecological niche modelling.  The fundamental niches
-are defined as [multivariate normal distributions](https://en.wikipedia.org/wiki/Multivariate_normal_distribution), 
-and the realised niches are created by defining the positive and negative interactions amongst a community of fundamental niches.  A full 
-description of the `virtualNicheR` package, including connections to niche theory and an explanation of the underlying mathematics, can be 
-found in the associated software paper [link to be included once published].
+To evaluate how well species-environment models perform in different situations, modellers need the ability to generate virtual niches that have known properties.  These virtual niches can then form the basis of virtual ecology experiments that can test how well different models perform under different conditions.
+
+`virtualNicheR` is an `R` package that is unique in that it provides an integrated framework that is needed to generate both [fundamental and realised niches](https://en.wikipedia.org/wiki/Ecological_niche).  We envisage that the ability to generate integrated fundamental and realised virtual niches will allow modellers to improve the virtual ecology experiments that have been advocated for evaluating species-environment relationship models.
+
+The fundamental niches are defined as [multivariate normal distributions](https://en.wikipedia.org/wiki/Multivariate_normal_distribution), and the realised niches are created by defining the positive and negative interactions amongst a community of fundamental niches.  A full description of the `virtualNicheR` package, including connections to niche theory and an explanation of the underlying mathematics, can be found in the associated software paper [link to be included once published].
 
 ## Citation
 
@@ -29,6 +28,9 @@ If you use `virtualNicheR` in your research we would be very grateful if you cou
 
 Etherington TR, Omondiagbe, OP (2019) virtualNicheR: creating virtual fundamental and realised niches [details to be updated once published]
 
+## Community guidelines
+
+We very much welcome input from others!  If you find a bug, need some help, or can think of some extra functionality that would be useful, please raise an [issue](https://github.com/manaakiwhenua/virtualNicheR/issues).  Better still, please feel free to fork the project and raise a pull request if you think and can fix a bug, clarify the documentation, or improve the functionality yourself.
 
 ## Installation
 
