### 2019-08-29 #1661 @tretherington
Many thanks @ViniciusBRodrigues and @lsw5077 for your comments 😃 !

I've updated things such that:

1. I have included a [Community guidelines](https://github.com/manaakiwhenua/virtualNicheR#community-guidelines) section in the README so folks know how best to report bugs, seek help, and suggest/make developments.  Apologies from me for neglecting to include that first time around 😞
2. I have reworded why the software is needed in the paper to try and make the need stronger and clearer.  You were both right that we had neglected to include that need in the README, so that has now been added there too.
3. I've also inserted what I think is the best easily accessed general reference to species-environment models in the paper so anyone interested to know more about their importance can look into things further.

@arfon I've also got the excellent whedon robot to regenerate the pdf, so hopefully everything should be up to date for you
