### 2019-09-25 #1661 @tretherington
@arfon I've created a new release v1.0.1 and I have archived the software in my institution's data store here: [https://datastore.landcareresearch.co.nz/dataset/virtualnicher](https://datastore.landcareresearch.co.nz/dataset/virtualnicher) which the DOI of [10.7931/cae9-bt94](https://doi.org/10.7931/cae9-bt94)

Do let me know if you need anything else! 😃
